pyproject.toml
[project]
name = "hr-pulse-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "azure-ai-textanalytics>=5.3.0",
    "azure-core>=1.38.1",
    "azure-identity>=1.25.2",
    "fastapi>=0.129.0",
    "joblib>=1.5.3",
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "opentelemetry-api>=1.39.1",
    "opentelemetry-exporter-otlp>=1.39.1",
    "opentelemetry-instrumentation-fastapi>=0.60b1",
    "opentelemetry-sdk>=1.39.1",
    "pandas>=3.0.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.0",
    "pyodbc>=5.3.0",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "sqlalchemy>=2.0.46",
    "streamlit>=1.9.0",
    "uvicorn[standard]>=0.41.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "jupyter>=1.1.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.1",
]
